+++ packages/fedify/src/federation/middleware.ts
@@ -1946,6 +1946,28 @@ export class ContextImpl<TContextData> implements Context<TContextData> {
     return result;
   }
 
+  getActorKeysByUsage(identifier: string): Promise<{
+    publicKeys: CryptographicKey[];
+    assertionMethods: Multikey[];
+  }> {
+    return this.getActorKeyPairs(identifier)
+      .then((keys) => ({
+        publicKeys: keys.map((k) => k.cryptographicKey),
+        assertionMethods: keys.map((k) => k.multikey),
+      }));
+  }
+
+  getActorFirstKeyByUsage(identifier: string): Promise<{
+    publicKey: CryptographicKey;
+    assertionMethod: Multikey;
+  }> {
+    return this.getActorKeysByUsage(identifier)
+      .then(({ publicKeys, assertionMethods }) => ({
+        publicKey: publicKeys[0],
+        assertionMethod: assertionMethods[0],
+      }));
+  }
+
   protected async getKeyPairsFromIdentifier(
     identifier: string,
   ): Promise<(CryptoKeyPair & { keyId: URL })[]> {